Fixed the failing unit tests

+++ test/stubby_integration/kvstore_integration_test.go
@@ -1,21 +1,22 @@
 package stubbyintegration
 
-
 import (
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
+// Stubby test for GetCollectionStats() kvstore service endpoint
 func TestGetCollectionStats(t *testing.T) {
 	result, err := getClient(t).KVStoreService.GetCollectionStats("namespace1", "collection1")
 	assert.Empty(t, err)
 	assert.NotEmpty(t, result)
 
-	assert.Equal(t, 5, result.Count)
+	assert.Equal(t, int64(5), result.Count)
 	assert.Equal(t, "namespace1", result.Ns)
-	assert.Equal(t, 1, result.Nindexes)
+	assert.Equal(t, int64(1), result.Nindexes)
 }
 
+// Stubby test for ping/GetServiceHealthStatus() kvstore service endpoint
 func TestGetServiceHealthStatus(t *testing.T) {
 	result, err := getClient(t).KVStoreService.GetServiceHealthStatus()
 	assert.Empty(t, err)
